update & clean
