+++ bot_command.py
@@ -1,7 +1,7 @@
 import re
 
 from jinja2 import Template
-from praw.models import Comment, Message, Redditor
+from praw.models import Comment, Message
 
 import bot_logger
 import config
@@ -348,48 +348,3 @@ def donate(reddit, msg, tx_queue, failover_time):
         bot_logger.logger.info('user %s not registered (command : donate) ' % user.username)
         msg.reply(Template(lang.message_need_register + lang.message_footer).render(username=user.username))
 
-
-# buy an reddit gold account for one month
-def gold(reddit, msg, tx_queue, failover_time):
-    user = models.User(msg.author.name)
-    if user.is_registered():
-        if msg.body.strip() == 'buy':
-
-            # todo : check if we have enough credits
-            gold_month = 1
-            if not gold_month >= 1:
-                # todo : store in db want an gold, when bot have new credits a PM can be send
-                msg.reply(Template(lang.message_gold_no_more).render(username=user.username))
-                return False
-
-            # check user confirmed balance is ok
-            if user.get_balance_confirmed() >= config.price_reddit_gold:
-                msg.reply(Template(lang.message_gold_no_enough_doge).render(username=user.username))
-                return False
-
-            # send amount of one month of gold to address
-            tx_id = crypto.tip_user(user.address, config.gold_address, config.price_reddit_gold, tx_queue,
-                                    failover_time)
-
-            if tx_id:
-                # send gold reddit
-                Redditor(reddit, user.username).gild(months=1)
-
-                # todo : update gold reddit table
-
-                # update user history
-                history.add_to_history(user, sender=user.username, receiver="Reddit", amount=config.price_reddit_gold,
-                                       action="buy reddit gold")
-
-                # send succes message
-                msg.reply(Template(lang.message_buy_gold_success).render(username=user.username))
-            else:
-                # send error message
-                msg.reply(Template(lang.message_buy_gold_error).render(username=user.username))
-
-        else:
-            # send info on reddit gold
-            msg.reply(Template(lang.message_buy_gold).render(username=user.username))
-    else:
-        bot_logger.logger.info('user %s not registered (command : donate) ' % user.username)
-        msg.reply(Template(lang.message_need_register + lang.message_footer).render(username=user.username))
